Refactor to extend from base class
