+++ includes/payment-gateway/class-card.php
@@ -1,0 +1,18 @@
+<?php
+/**
+ * Class Sepa
+ *
+ * @package WCPay\Payment_Gateway
+ */
+
+namespace WCPay\Payment_Gateway;
+
+use WC_Payment_Gateway_WCPay;
+
+/**
+ * Credit Card Payment method.
+ * Right now behaves exactly like WC_Payment_Gateway_WCPay for max compatibility.
+ */
+class Card extends WC_Payment_Gateway_WCPay {
+
+}
